refactor: move appDataDir to DataDirectory file
Refs: #381
