+++ lib/lokksmith-core/src/jvmMain/kotlin/dev/lokksmith/DataDirectory.kt
@@ -30,3 +30,22 @@ public sealed interface DataDirectory {
     /** Uses a custom absolute path as the data directory. */
     public data class Custom(val path: File) : DataDirectory
 }
+
+internal fun appDataDir(appName: String): File {
+    val userHome = System.getProperty(USER_HOME_PROPERTY)
+    return when (OperatingSystem.current) {
+        OperatingSystem.Windows -> File(System.getenv(WINDOWS_APPDATA_ENV), appName)
+        OperatingSystem.MacOS -> File(userHome, "$MACOS_APP_SUPPORT_PATH/$appName")
+        OperatingSystem.Linux ->
+            File(
+                System.getenv(LINUX_XDG_DATA_HOME_ENV) ?: "$userHome/$LINUX_LOCAL_SHARE_PATH",
+                appName,
+            )
+    }
+}
+
+private const val USER_HOME_PROPERTY = "user.home"
+private const val WINDOWS_APPDATA_ENV = "APPDATA"
+private const val MACOS_APP_SUPPORT_PATH = "Library/Application Support"
+private const val LINUX_XDG_DATA_HOME_ENV = "XDG_DATA_HOME"
+private const val LINUX_LOCAL_SHARE_PATH = ".local/share"
